Revert "Merge branch 'mbqcfication' of https://github.com/CQCL/tket into mbqcfication"
This reverts commit 907306918d4db976242f6da79546424c53fdaf22, reversing
changes made to af78fd7e629c2f8dc3e09d6a16084e2191fc479b.
